refactor: NetworkManager 리팩토링(제네릭 타입 사용 추가), RequestGenerator 분리
