+++ ChatBot/ChatBot/Models/AIAnswerModel.swift
@@ -15,7 +15,7 @@ struct AIAnswerModel: Decodable {
     let model: String
     let choices: [Choice]
     let usage: Usage
-    let systemFingerprint: String
+    let systemFingerprint: String?
     
     enum CodingKeys: String, CodingKey {
         case id, object, created, model, choices, usage
